# fix: reduce allocations in the Time Window Block Selector